Implement background navigation tracker
This keeps track of the various lolesports tabs and the urls navigated
to. This information is sent to the content scripts so that they can
analyse whether to connect or disconnect any mutation observers.

Also, due to the way lolesports changes links, content scripts are now
added to the whole website rather than select paths.

The changes to the manifest also deal with the situations, where there's a
subdomain like: watch.euw.lolesports.com. Although, at the time of
writing, the website doesn't use this format. It has been used before though.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>