Break up spoiler.js into modules
This is necessary for two reasons:
1. The file was getting large and crowded.
2. Some of the functionality would be required by other content scripts
and copy pasting the code wouldn't be ideal

Also, as part of this change, the info class had to be generalised as it
will be used in other content scripts to come.

Removed the unused function in utils.js. This will be added if and when
necessary.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>