Refactor tracking of tab links
It was realised that using the local storage as an 'event' mechanism
would not scale well beyond 1 content script. Removing this mechanic
and opting for a simpler one wouldn't work with the Info class.

The use of closures ended up being as the best alternative as it maintains
the goals of the Info class, which was to keep track of the links the
user has visited and is currently on.

As a result of this refactoring, sending the tabId from the background script
was no longer needed. Also, the utils file was rendered useless as none of
its functions were in use.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>