Refactor sidebar.js
Since the click listeners for the left and right sidebar options were
very similar, it made sense to be able to simplify them into a single
function.

Using closures the above goal is achieved. Also the closure is used
to store the state of the size of the left sidebar. This is in anticipation
of various screen resolutions having separate sizes.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>