Fix switching between all 3 layout options
A major problem that this commit fixes is that switching between the
right layout to theatre wasn't working.

The previous code also had a bug where it was trying to access
elements which disappeared after the options button is closed.

Another bug was that the event listeners for the left sidebar and theatre
were being added and called more than once. This resulted in unintended
UI changes.

Refactored a majority of the functions to make them easier to maintain.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>