Change project structure
With this structure, it'll make things easier when time comes to porting to
Chrome. Also, the structure makes building for either browser much easier.

Part of this fix includes a hack where:
To trigger the copy in rollup during watch mode, the spoilers.js file has
to be saved in order to trigger a copy. This is because at the time of
writing, rollup-plugin-copy doesn't support the watch mode.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>