refactor StatsInfo
Having the class in its own file allows for much better flexibility in
functionality between live and vods.

Also the checking of files and loading of audio files are now part of
the setup. This makes much more sense than having them in init.

Having the statsInfo object being created in init is better because
before it was being created whether or not the user had enabled
voice announcers.

Deprecated methods and variables that will soon be rendered useless
both for vods and live matches.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>