Change selection of announcer ally team
The previous system of clicking on the team name to choose which
one would be considered the ally team by the announcer was riddled
with bugs and was unreliable.

Now, the selection is among the options in the "options" menu on the bottom
right of the video (alongside layout and player options). This is much
better for the user. Plus it is much more reliable now.

There's no longer a need to have a notification to indicate
which team the user has selected. The menu provides enough visual
feedback.

These changes include the refactoring of stats.js to align it with the
refactoring done to mutation.js

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>