Update build tools
This simplifies the rollup config file and allows for more flexibility
and scalability when adding more javascript files and plugins.

To achieve this goal the copying of static files is now delegated to
make. Make was chosen since it is language agnostic and it allows
for one build process to be used whether on windows, linux or on
the CI server. Also, the rules on make only copy static files that have
changed, which is more efficient than the rollup-plugin-copy.

This change made the rollup-plugin-copy unneeded. Inorder to remove
it as a dependency, a package.json had to be created.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>