Add JsDoc comments
This is to help with static type checking using the typescript compiler
in VS Code. Also, several areas of the code desparately needed to be
documented in order to make understanding them easier.

To prevent having the jsdoc type comments making a lot of code noise
it is now preferable to define variables as const and only use let for
variables whose values change. Since most types can be inferred, the
variables defined with let are the only ones which may need explicit
jsDoc type comments. The eslint config is updated to enforce this
change

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>