Improve makefile
The makefile previously contained a separate rule for copying the
manifest file into the build directory. This was redundant since the
rule was similar to the other one for the other static files. This same
logic applies to the rule that created the build directory.

Also, utilised variables in defining the rule containing static patterns

Lastly, a new phony target that does cleanup was added.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>