Update build instructions in README.md
The previous instructions did not take into account the changes made in
the build process that introduced the use of make. Also, with rollup being
added as a dev dependency, the instructions are no longer valid.

To provide a much cleaner build instruction, two make targets were added
in the Makefile: clean_install and clean_build. clean_install uses the
command: npm ci to do a fresh installation of the dependencies.
clean_build creates a new build of the extension.

Signed-off-by: Victor Otieno <vickz84259@gmail.com>