Improve makefile
The primary goal of these changes was to make sure that updating a js
file only cause rollup to run for that specific file. Previously it would
update every file even unnecessary ones.

Also included is the creation of a zip file since both Firefox and Chrome
allow for extensions to be packaged in zip files.

Signed-off-by: Victor Otieno <vickz84259@gmail.com>