Add tsconfig.json
This represents the configurations for the Typescript compiler. The
allowJs option is temporary. It will be removed once the whole project has
been fully migrated to TypeScript. Similarly, the noEmitOnError option
has been set to false to assist with the transition. Later it will be set to
true.

Also, the removeComments option currently doesn't do what it is
intended to do. Once the transition is complete the decision of whether
to continue using the rollup-cleanup plugin will be made.

Signed-off-by: Victor Otieno <vickz84259@gmail.com>